Use the hchr field for Formaldehyde values (#15)
Reference discussion link: https://github.com/libdyson-wg/ha-dyson/issues/4

The `hcho` field in MQTT is of an unknown unit/scale and has no known
conversion factor to a known unit. However, it was recently discovered
that `hchr` (referred to as "High Res Formaldehyde" in the android app)
has a clean conversion factor of n/1000 to mg/m^3.